requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
Pillow==10.1.0
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.0
spacy==3.7.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
psycopg2-binary==2.9.9
google-generativeai==0.3.2
django-storages==1.14.2
django-anymail==10.2
nltk==3.8.1